requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.9
Flask-Cors==3.0.3
psycopg2-binary==2.8.2
pymongo==3.6.0
python-dotenv==0.7.1
schema==0.6.7
bcrypt==3.1.4
sqlalchemy==1.3.3
eventlet==0.25.2
gunicorn==20.0.4